컨트롤러 (Shelter, Volunteer) -> IPS 으로 통합
